+++ src/globals/consts.lua
@@ -6,5 +6,6 @@ return {
     GROUND_H = 50,
     FONT_TITLE_SIZE = 80,
     FONT_HEADER_SIZE = 48,
+    FONT_MAIN_SIZE = 36,
     FONT_SUB_SIZE = 16,
 }